+++ src/main/java/fr/maxlego08/shop/buttons/ZShowConfirmItemButton.java
@@ -1,5 +1,6 @@
 package fr.maxlego08.shop.buttons;
 
+import fr.maxlego08.menu.api.utils.Placeholders;
 import fr.maxlego08.shop.ShopPlugin;
 import fr.maxlego08.shop.api.PlayerCache;
 import fr.maxlego08.shop.api.buttons.EconomyAction;
@@ -7,6 +8,7 @@ import fr.maxlego08.shop.api.buttons.ShowConfirmItemButton;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.Plugin;
+import org.jetbrains.annotations.NotNull;
 
 public class ZShowConfirmItemButton extends ShowConfirmItemButton {
 
@@ -17,12 +19,12 @@ public class ZShowConfirmItemButton extends ShowConfirmItemButton {
     }
 
     @Override
-    public ItemStack getCustomItemStack(Player player) {
+    public ItemStack getCustomItemStack(@NotNull Player player, @NotNull Placeholders placeholders) {
 
         PlayerCache playerCache = this.plugin.getShopManager().getCache(player);
         EconomyAction itemButton = playerCache.getEconomyAction();
-        if (itemButton == null) return super.getCustomItemStack(player);
+        if (itemButton == null) return super.getCustomItemStack(player, placeholders);
 
-        return itemButton.getCustomItemStack(player);
+        return itemButton.getCustomItemStack(player, placeholders);
     }
 }