+++ src/main/java/fr/maxlego08/shop/buttons/ZShowItemButton.java
@@ -1,5 +1,6 @@
 package fr.maxlego08.shop.buttons;
 
+import fr.maxlego08.menu.api.utils.Placeholders;
 import fr.maxlego08.shop.ShopPlugin;
 import fr.maxlego08.shop.api.PlayerCache;
 import fr.maxlego08.shop.api.buttons.ItemButton;
@@ -8,6 +9,7 @@ import fr.maxlego08.shop.placeholder.Placeholder;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,13 +25,13 @@ public class ZShowItemButton extends ShowItemButton {
     }
 
     @Override
-    public ItemStack getCustomItemStack(Player player) {
+    public ItemStack getCustomItemStack(@NotNull Player player, @NotNull Placeholders placeholders) {
 
         PlayerCache playerCache = this.plugin.getShopManager().getCache(player);
         ItemButton itemButton = playerCache.getItemButton();
-        if (itemButton == null) return super.getCustomItemStack(player);
+        if (itemButton == null) return super.getCustomItemStack(player, placeholders);
 
-        ItemStack itemStack = itemButton.getCustomItemStack(player);
+        ItemStack itemStack = itemButton.getCustomItemStack(player, placeholders);
         itemStack.setAmount(playerCache.getAmount());
 
         ItemMeta itemMeta = itemStack.getItemMeta();